Move license from README to new LICENSE doc file
One technical problem with the existing license is that the copyright
notice has not been updated, either with later years in which changes
were made, or notices as to any copyrights held by contributors not
employed by the original (and only) copyright holder listed.

Given that non-compliance with the chosen license, I suspect [the
Unlicense](https://unlicense.org/) is more appropriate.